yocto_qemu: add WIP metadata
This is the beginning of a labgrid environment configuration for making
use of QEMU on a system to boot a Yocto-produced image. More
specifically, this is tailored to try booting a
core-image-ptest-python3-numpy image produced for qemuriscv64. The
details of what the Yocto-based runqemu tool actually runs when 'runqemu
nographic snapshot' is invoked are stored in qemu.log as a reference,
and the lg-qemu-env.yaml file is based heavily on the one in
labgrid/examples/qemunetworking/, but it tries to match the runqemu
arguments as closely as possible. qemunetworkstrategy.py is currently
unchanged from the one in the examples directory.

This doesn't quite work right now - with 'display: qemu-default' and
without '-nographic' in extra_args, the QEMU graphical window will open
and the VM will spin up, but no actual output is seen. This may be an
issue with the -serial options provided, or there may be something else
going on, because after a long enough duration the VM exits on its own.

Signed-off-by: Trevor Gamblin <tgamblin@baylibre.com>